+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log (vscode-powertools)
 
+## 0.25.0 (February 18th, 2019; code execution)
+
+* added following constants for [Code Execution](https://github.com/egodigital/vscode-powertools/wiki/Tools#code-execution):
+  * `$cb`: copyies a value to clipboard
+  * `$cmyk`: converts CMYK color from or to hex
+  * `$hsl`: converts HSL color from or to hex
+  * `$rbg`: converts RGB color from or to hex
+
 ## 0.24.0 (February 18th, 2019; Azure DevOps)
 
 * can open Azure DevOps dashboards now via `Power Tools: Azure DevOps` command